Backend - MOPI - Restaurante/requirements_updated.txt
asgiref==3.9.2
Brotli==1.1.0
dj-database-url==3.0.1
Django==5.2.7
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
pillow==11.3.0
whitenoise==6.11.0
django-jazzmin==3.0.0
django-cors-headers==4.3.1
